Increase robustiness in resource-constrained environments
This change allows doks-debug to operate in environments where resources
are partially constrained, e.g., when nodes are under disk pressure. We
do so by turning the pod into a guaranteed one and setting a priority
class applicable in DOKS. The latter also requires to run in the
kube-system namespace, which is likely a good measure in general.